requirements.txt
streamlit
psycopg2-binary
pydantic
plotly
python-dotenv
google-generativeai
PyYAML

# Development and code quality
pre-commit
black
ruff
